Increase the accuracy of the position fields form Float to Double
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20417 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: e7d36a63d38ebb916f896e03cec22e33422d855a

+++ server/wrapper/nmea/de.uniol.inf.is.odysseus.wrapper.nmea/src/de/uniol/inf/is/odysseus/wrapper/nmea/sentence/aissentences/payload/decoded/PositionReport.java
@@ -26,8 +26,8 @@ public abstract class PositionReport extends DecodedAISPayload {
 	private final Integer rateOfTurn;
 	private final Float speedOverGround;
 	private final Boolean positionAccurate;
-	private final Float latitude;
-	private final Float longitude;
+	private final Double latitude;
+	private final Double longitude;
 	private final Float courseOverGround;
 	private final Integer trueHeading;
 	private final Integer second;
@@ -37,8 +37,8 @@ public abstract class PositionReport extends DecodedAISPayload {
 	protected PositionReport(String originalNmea, AISMessageType messageType,
 			Integer repeatIndicator, MMSI sourceMmsi,
 			NavigationStatus navigationStatus, Integer rateOfTurn,
-			Float speedOverGround, Boolean positionAccurate, Float latitude,
-			Float longitude, Float courseOverGround, Integer trueHeading,
+			Float speedOverGround, Boolean positionAccurate, Double latitude,
+			Double longitude, Float courseOverGround, Integer trueHeading,
 			Integer second, ManeuverIndicator maneuverIndicator,
 			Boolean raimFlag) {
 		super(originalNmea, messageType, repeatIndicator, sourceMmsi);
@@ -71,11 +71,11 @@ public abstract class PositionReport extends DecodedAISPayload {
 		return positionAccurate;
 	}
 
-	public final Float getLatitude() {
+	public final Double getLatitude() {
 		return latitude;
 	}
 
-	public final Float getLongitude() {
+	public final Double getLongitude() {
 		return longitude;
 	}
